% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_egnyte.R
\name{authenticate_egnyte}
\alias{authenticate_egnyte}
\title{Authenticate a user using the Egnyte Filesystem API}
\usage{
authenticate_egnyte(username, password, client_id, domain)
}
\arguments{
\item{client_id}{The default client ID for bulldriver. Can leave as default unless using different API.}

\item{RB_username}{Your Red Bull corporate username}

\item{RB_password}{Your Red Bull corporate password}

\item{autoset_env}{If left as TRUE will save the resulting token as an environment variable}

\item{url}{The top-level Egnyte domain}
}
\description{
This function authenticates a user with the \href{https://developers.egnyte.com/docs/read/Public_API_Authentication}{Egnyte Filesystem API}.
}
\details{
In order for this function to run you must first create
a new API key for your application via the \href{https://developers.egnyte.com/member/register}{Egnyte for Developers} site.
}
