% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_user.R
\name{authenticate_user}
\alias{authenticate_user}
\title{Authenticate a user using the Egnyte Filesystem API}
\usage{
authenticate_user(domain, app_key, username, password)
}
\arguments{
\item{domain}{the URL of your Egnyte domain}

\item{app_key}{API key of your Egnyte application}

\item{username}{user's Egnyte username}

\item{password}{user's Egnyte password}
}
\description{
This function authenticates a user for use of an \href{https://developers.egnyte.com/docs/read/Public_API_Authentication}{Egnyte Filesystem API} application.
In order for this function to work you (or an admin) must first create a new API key for your application via the \href{https://developers.egnyte.com/member/register}{Egnyte for Developers} site.
}
\section{Details}{


Use this function to authenticate a specific user for your Egnyter API application.
This function will return the authorisation token that is required for the Egnyter \code{upload_*} and \code{download_*}
functions to operate.

You can run this function directly every time your require an authorisation token or use \code{\link{set_token}} to save
the token to your options.

If your Egnyte username or password changes you will need to re-authenticate yourself and get a new token to continue using Egnyter.
}

