% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{create_folder}
\alias{create_folder}
\title{Create a new folder in an Egnyte filesystem}
\usage{
create_folder(parent_folder, new_folder_name,
  token = get_parameter("token"), domain = get_parameter("domain"))
}
\arguments{
\item{parent_folder}{The full path (starting 'Shared/') of the location that the new folder will be created}

\item{new_folder_name}{The name of the new folder to create}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}
}
\description{
This function will create a new folder/directory in an Egnyte filesystem
}
