% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_delim}
\alias{download_delim}
\title{Download a delimited file from Egnyte into a local data frame}
\usage{
download_delim(source, delim, encoding = "ISO-8859-1",
  token = get_parameter("token"), domain = get_parameter("domain"),
  ...)
}
\arguments{
\item{source}{The path of the remote file in your Egnyte filesystem}

\item{delim}{Single character used to separate fields within a record}

\item{encoding}{The default encoding to use for content translation}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}

\item{...}{Additional arguments to be passed to \code{read_delim}}
}
\description{
This function downloads and parses a delimited file from Egnyte into
a local data frame in R.
}
