% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download a file from Bulldrive (any format)}
\usage{
download_file(file_path, file_type = "text",
  token = Sys.getenv("EGNYTETOKEN"), url = Sys.getenv("EGNYTEDOMAIN"),
  encoding = "ISO-8859-1")
}
\arguments{
\item{file_path}{The full path (starting 'Shared/') of the remote file}

\item{file_type}{The MIME file type to pull the file in. See \link[httr]{content} for more info}

\item{token}{Your Bulldrive authentication token (create with \link[bulldriver]{get_token})}

\item{url}{The top-level Egnyte domain}

\item{encoding}{The default encoding to use for the content translation}
}
\description{
This function downloads and parses a file of any type from Bulldrive.
The raw content will be returned and will need to be parsed for further
use if needed.
}
