% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_rdata}
\alias{download_rdata}
\title{Download an RData file from Egnyte and load contents into R}
\usage{
download_rdata(source, encoding = "ISO-8859-1",
  token = get_parameter("token"), domain = get_parameter("domain"),
  ...)
}
\arguments{
\item{source}{The path of the remote file in your Egnyte filesystem}

\item{encoding}{The default encoding to use for content translation}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}

\item{...}{Additional arguments to be passed to \code{load}}
}
\description{
This function downloads and parses an RData file from Egnyte into
local objects (as if you had loaded the file using \link[base]{load}).
}
