% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_file.R
\name{list_file}
\alias{list_file}
\title{List file or folder}
\usage{
list_file(
  source,
  token = get_parameter("token"),
  domain = get_parameter("domain")
)
}
\arguments{
\item{source}{The full path (starting 'Shared/') of the remote file or folder}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}
}
\description{
List information about a file or folder
}
