% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_csv}
\alias{upload_csv}
\title{Upload a csv file to Egnyte}
\usage{
upload_csv(x, dest, token = egnyter::get_parameter("token"),
  domain = egnyter::get_parameter("domain"), ...)
}
\arguments{
\item{x}{A data frame to upload to Egnyte}

\item{dest}{The remote Egnyte file path you want to upload to}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}

\item{...}{Additional arguments to be passed to \code{write_csv}}
}
\description{
This function converts a local data frame into a csv file and uploads it to a specified Egnyte directory.
Data frame to csv file conversion is done using \code{\link[readr]{write_csv}}.
}
