% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_feather}
\alias{upload_feather}
\title{Upload a feather file to Egnyte}
\usage{
upload_feather(x, dest, token = egnyter::get_parameter("token"),
  domain = egnyter::get_parameter("domain"), ...)
}
\arguments{
\item{x}{A data frame to upload to Egnyte}

\item{dest}{The remote Egnyte folder you want to upload to}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}

\item{...}{Additional arguments to be passed to \code{write_feather}}
}
\description{
This function converts a local data frame into a feather file and uploads it to a specified Egnyte directory.
Data frame to feather file conversion is done using \code{\link[feather]{write_feather}}.
}
