% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Download a file from Egnyte}
\usage{
download_file(source, local_dest, file_type = "raw",
  encoding = "ISO-8859-1", token = get_parameter("token"),
  domain = get_parameter("domain"))
}
\arguments{
\item{source}{The full path (starting 'Shared/') of the remote file}

\item{local_dest}{The local file to be created}

\item{file_type}{The MIME file type to pull the file in, see \link[httr]{content} for more info}

\item{encoding}{The default encoding to use for the content translation, see \link[httr]{content} for more info}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}
}
\description{
This function downloads and parses a file of any type from Egnyte.
}
