% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{set_token}
\alias{set_token}
\title{Save authorisation details for an Egnyter user}
\usage{
set_token(domain, app_key, username, password)
}
\arguments{
\item{domain}{the URL of your Egnyte domain}

\item{app_key}{API key of your Egnyte application}

\item{username}{user's Egnyte username}

\item{password}{user's Egnyte password}
}
\description{
This function allows you to save your Egnyter authorisation token
(which you can get using \code{\link{authenticate_user}}) and Egnyte domain
to your R options which means you will not have to reference them every time you
use Egnyter.
}
