% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_rdata}
\alias{upload_rdata}
\title{Upload an RData file to Egnyte}
\usage{
upload_rdata(..., dest, token = egnyter::get_parameter("token"),
  domain = egnyter::get_parameter("domain"))
}
\arguments{
\item{...}{The names of the objects to be saved}

\item{dest}{The remote Egnyte file path you want to upload to}

\item{token}{User's Egnyter authorisation token}

\item{domain}{Egnyte domain URL}
}
\description{
This function saves a local object as an RData file and uploads it to a specified Egnyte directory.
Object to RData conversion is done using \code{\link[base]{save}}.
}
