% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{upload_rdata}
\alias{upload_rdata}
\title{Upload an RData file to Egnyte}
\usage{
upload_rdata(x, path, token = egnyter::get_parameter("token"),
  domain = egnyter::get_parameter("domain"))
}
\arguments{
\item{x}{The full path (starting 'Shared/') of the remote file}

\item{path}{The MIME file type to pull the file in. See \link\link{httr}{content} for more info}

\item{token}{Your Bulldrive authentication token (create with \link\link{bulldriver}{get_token})}

\item{domain}{The top-level Egnyte domain}
}
\description{
This function downloads and parses a file of any type from Bulldrive.
The raw content will be returned and will need to be parsed for further
use if needed.
}
